Make resolveConfig compatible with feature flag configs (#2347)
* Make resolveConfig compatible with feature flag configs

* Update changelog
