+++ resolveConfig.js
@@ -1,6 +1,11 @@
 const resolveConfigObjects = require('./lib/util/resolveConfig').default
-const defaultConfig = require('./stubs/defaultConfig.stub.js')
+const getAllConfigs = require('./lib/util/getAllConfigs').default
 
 module.exports = function resolveConfig(...configs) {
-  return resolveConfigObjects([...configs, defaultConfig])
+  // Make sure the correct config object is mutated to include flagged config plugins.
+  // This sucks, refactor soon.
+  const firstConfigWithPlugins = configs.find(c => Array.isArray(c.plugins)) || configs[0]
+  const [, ...defaultConfigs] = getAllConfigs(firstConfigWithPlugins)
+
+  return resolveConfigObjects([...configs, ...defaultConfigs])
 }
